[Rename, Create] - ViewController를 MainView와 MainViewController로 분리
